# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API UHS Asbl",
        "description": "API REST complète avec gestion des activites de UHS Asbl",
        "contact": {
            "name": "Support API",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api-uhs.uhs-rdc.org",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/facturations.getAllData": {
            "get": {
                "tags": [
                    "Facturations"
                ],
                "summary": "Liste des facturations",
                "description": "Récupérer toutes les facturations avec leurs abonnés",
                "operationId": "92837833ecfbbd98d604b56990a0be60",
                "responses": {
                    "200": {
                        "description": "Liste des facturations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Facturation"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/facturations.delete/{id}": {
            "delete": {
                "tags": [
                    "Facturations"
                ],
                "summary": "Supprimer une facturation",
                "description": "Suppression d’une facturation par ID",
                "operationId": "3c4f8e663d9a08ecf825ec5f71ac3dee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la facturation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Facturation supprimée avec succès"
                    },
                    "404": {
                        "description": "Facturation non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/facturations.store": {
            "post": {
                "tags": [
                    "Facturations"
                ],
                "summary": "Créer une facturation",
                "description": "Ajout d’une nouvelle facturation pour les abonnés",
                "operationId": "6bb0c070c179910d08e0ac69ad108cd5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date_emission"
                                ],
                                "properties": {
                                    "date_emission": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Facturation créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Facturation ajoutée avec succès"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Facturation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/user.delete/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a User",
                "description": "Delete a user by their ID.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Facturation": {
                "title": "Facturation",
                "description": "Modèle de facturation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "abonne_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "addedBy": {
                        "type": "integer",
                        "example": 1
                    },
                    "mois": {
                        "type": "string",
                        "example": "09-2025"
                    },
                    "montant": {
                        "type": "number",
                        "format": "float",
                        "example": 150.75
                    },
                    "dette": {
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "status": {
                        "type": "string",
                        "example": "Non payé"
                    },
                    "reference": {
                        "type": "string",
                        "example": "FAC-6637"
                    },
                    "date_emission": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-09-01"
                    },
                    "date_paiement": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-09-10",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-21T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-21T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}